+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 - Add suppot for suppressing OK output lines
 - Updated README to include information on annotations and check token substitution
+- If a search string is "" then skip it, do not fail it
 
 ## [0.1.0] - 2020-11-04